Correct, modernize and make more independent of user
The playbook had drifted away from the text of `README.md`, so I corrected what I could.

I eliminated literal user name.

I modified the instructions so the `credentials_file` name is constant.

I modified the instructions so that the current value of `project` in `gce_vars/auth` need not be
`YOUR_PROJECT_ID`. This is because it is too easy to forget to change this value from an
actual project ID to `YOUR_PROJECT_ID` before committing.

Replaced deprecated `include` with `import_playbook`.

There is no file `clean-up.yml`, so I corrected it to `cleanup.yml`. Although I did not say or do
anything about it, `cleanup.yml` is quite broken.